requirements.txt
-i https://pypi.org/simple
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-python-client==1.12.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-auth==1.22.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httplib2==0.18.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
lxml==4.6.1
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pytz==2020.1
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
selenium==3.141.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
